Cleanly implement DDGProxy and reflect
