+++ .gitignore
@@ -9,6 +9,7 @@ public/js/*.*
 public/css/*.*
 *.swp
 *.swo
+build/fingerprint.js
 build/firefox
 build/chrome
 build/safari
