+++ package.json
@@ -6,7 +6,7 @@
     "repository": "duckduckgo/duckduckgo-privacy-extension",
     "scripts": {
         "postinstall": "npm run copy-sjcl && export PATH='$PATH:$PWD/node_modules/chromedriver/bin'",
-        "copy-sjcl": "cd node_modules/sjcl/ && ./configure --no-export --compress=none --without-all --with-hmac --with-codecHex && make && cp sjcl.js ../../shared/js/content-scope/",
+        "copy-sjcl": "cd node_modules/sjcl/ && ./configure --compress=none --without-all --with-hmac --with-codecHex && make && cp sjcl.js ../../shared/js/content-scope/",
         "build": "echo 'Try npm run release-firefox or npm run release-chrome instead' && exit 0",
         "dev": "echo 'Try npm run dev-firefox or npm run dev-chrome instead' && exit 0",
         "release": "grunt build && node scripts/httpsWhitelist.js",
