+++ shared/js/content-scope/audio.js
@@ -1,4 +1,4 @@
-import { shouldExemptMethod, iterateDataKey } from "./utils";
+import { shouldExemptMethod, iterateDataKey, DDGProxy, DDGReflect } from "./utils";
 
 export function initAudioProtection (args) {
     const { sessionKey, site } = args
