+++ shared/js/content-scope/canvas.js
@@ -1,4 +1,4 @@
-import { shouldExemptMethod, iterateDataKey } from "./utils";
+import { shouldExemptMethod, iterateDataKey, DDGProxy, DDGReflect } from "./utils";
 
 export function initCanvasProtection (args) {
     const { sessionKey, site } = args
